Improved code style
